apiloanspayments/requirements.txt
asgiref==3.7.2
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.2.0
cryptography==41.0.2
defusedxml==0.7.1
dj-rest-auth==4.0.1
Django==4.2.3
django-allauth==0.54.0
django-filter==23.2
django-rest-auth==0.9.5
django-upgrade==1.14.0
djangorestframework==3.14.0
idna==3.4
Markdown==3.4.3
oauthlib==3.2.2
pycparser==2.21
PyJWT==2.8.0
python3-openid==3.2.0
pytz==2023.3
requests==2.31.0
requests-oauthlib==1.3.1
six==1.16.0
sqlparse==0.4.4
tokenize-rt==5.1.0
typing_extensions==4.7.1
tzdata==2023.3
urllib3==2.0.4